requirements.txt
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.5
crayons==0.4.0
english-words==1.1.0
idna==3.4
imutils==0.5.4
iniconfig==1.1.1
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
piexif==1.1.3
Pillow==9.2.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pytesseract==0.3.10
pytest==7.1.3
pytest-socket==0.5.1
requests==2.28.1
stegano==0.10.2
tomli==2.0.1
urllib3==1.26.12
wordninja==2.0.0
